Skip account-level WorkspaceIT test in azure (#55)
## Changes
Account-level workspace administration is done through ARM APIs instead
of the Databricks SDK. This PR makes two changes:

1. Add the DisableIfEnvironmentVariable annotation to WorkspacesIT to
indicate that it should be disabled if the ARM_CLIENT_ID environment
variable is defined.
2. Add support to EnvTest to inspect for the
DisableIfEnvironmentVariable annotation is present on a test class. If
so, evaluate the logic for skipping environment variables, but check
against the local environment computed by EnvTest. If multiple
annotations are present, they all are checked, and if any one is
disabled, the overall test is disabled as well.

## Tests
- [x] WorkspacesIT does not run when an Azure environment is used.