Set necessary headers when authenticating via Azure CLI (#136)
## Changes
The Java SDK request authentication logic is inconsistent between the
Azure login types: for service principal auth, the SDK correctly adds
the X-Databricks-Azure-Workspace-Resource-Id when configured, but this
is missed for Azure CLI auth. Additionally, when logging in via Azure
CLI using a service principal, the service management token must also be
fetched from the CLI.

This PR fixes this by defining the logic to attach these header in a
common function that is used by all Azure-specific authentication types.

See https://github.com/databricks/databricks-sdk-go/pull/584 for the
same change in the Go SDK.
See https://github.com/databricks/databricks-sdk-py/pull/290 for the
same changes in the Python SDK.
## Tests
- [x] Unit tests to cover the two scenarios for Azure CLI w.r.t.
management endpoint token fetching, and one to verify that
X-Databricks-Azure-Workspace-Resource-Id is included when using Azure
CLI.