Support inputstream for request and response bodies (#138)
## Changes
This PR adds support for non-application/json requests and responses.
This PR is divided into two parts:

**Support setting headers per request.** The main methods of the
ApiClient interface are expanded to accept one new
parameter,`Map<String, String> headers`, and all callers are expected to
pass a map of headers to be included in the request. As the allowed
content types for requests and responses are known from the OpenAPI
specification, impls must construct the request header map in code
generation and pass it to the ApiClient. The default "Content-Type:
application/json" header is removed, as each request should specify its
own Content-Type and Accept headers. (This is merged in
https://github.com/databricks/databricks-sdk-java/pull/135.)

**Add support for streaming request and response bodies to support
non-application/json requests/responses.** Today, serialization of
requests is done in the ApiClient. This implies that ApiClient needs to
be able to serialize a request purely based on the parameters provided
to it via headers, the request body, etc. In this proposal, the
serialization is coupled to the request type: InputStream indicates that
the client has already serialized the request, so it can be provided
directly to the underlying HttpClient library, and any other type will
be passed through Jackson's ObjectMapper for serialization to JSON.
Internally, request and response bodies are modeled as InputStreams
rather than byte arrays to support streaming requests and responses,
though for non-streaming requests and responses, a string value for the
body is captured in the `debugBody` field of Request and Response,
respectively.

One important caveat about streaming responses is that callers are
required to close the streams when they are done using them. This
releases the underlying HTTP connection back to the connection pool for
use in subsequent requests. Otherwise, customers could encounter
deadlocks waiting for connections. For this, we recommend that users use
the try-with-resources model as demonstrated in integration test
FilesIT.java.

## Tests
- [x] Integration test for Files IT exercises uploading and downloading
files using the streaming pathways.
- [x] Existing integration tests cover the pre-existing non-streaming
pathways.