Fix query param serialization for requests with enums (#140)
## Changes
Query parameters are determined reflectively from request classes by
scanning for fields with a QueryParam annotation. Serialization of query
parameters is recursive in order to support complex types like the
filter structures used for listing in the SQL query history service.
This PR makes the following changes:

1. Terminate recursion when the request field is an enum.
2. When iterating through the request object's fields, skip any fields
not annotated with QueryParam (but recursively, all fields do need to be
serialized).
3. Rename the inner class from HeaderEntry to QueryParamPair (it
represents query param pairs, not header entries).

## Tests
Test for this change is dependent on other testing refactors that will
be merged as part of
https://github.com/databricks/databricks-sdk-java/pull/139.