Add ErrorInfo to API errors (#153)
## Changes
Databricks API can return error details in case of errors. In some
cases, users need to access such details to be able to solve the issue.
This is the case for the errors of type ErrorInfo used by the Settings
Platform.
This PR adds the necessary fields to the APIError to Unmarshal the
ErrorDetails and to expose the details of type ErrorInfo to the users.

## Tests
- [X] Unit test
- [X] `make fmt`
- [X] Integration test