Implement notebook-native auth for the Java SDK (#171)
## Changes
Notebook-native authentication has been supported in the Python SDK for
the last several months, and customers using the SDKs in Databricks
appreciate the easier CUJ for authentication. This PR ports the behavior
to the Java SDK. The approach taken for this implementation differs
somewhat from the Python SDK's implementation. In particular, in Python,
DBR provides some functionality to get the command context (`from
dbruntime.databricks_repl_context import get_context`, `from
dbruntime.sdk_credential_provider import init_runtime_native_auth`) as
well as from IPython. The Java SDK uses the DBUtils object directly.
However, the SDK doesn't depend on the DBUtils API (which is only in
Scala), so instead we load it reflectively.

## Tests
Ran in a notebook in DBR, worked.