Support custom scopes and redirectUrl for U2M OAuth flow (#190)
## Changes

Added following utility methods to DatabricksConfig

a. setRedirectUrl: This will be passed to OAuthClient and will be used
in U2M OAuth flow if client wants to provide a custom redirect Url as
per OAuth configuration of their application.
b. setScopes: Clients can set custom scopes for U2M OAuth flow. If
offline_access is not present, it will be added to the scopes for
refresh token support.

## Tests

Tested manually using JDBC driver as a client. Also, added test case for
custom values of redirectUrl and scopes.