Fix SCIM pagination (#193)
## Changes
Supersedes https://github.com/databricks/databricks-sdk-java/pull/188.

This PR incorporates two hard-coded changes for the SCIM API in the
Python SDK:

startIndex starts at 1 for SCIM APIs, not 0. However, the existing
.Pagination.Increment controls both the start index as well as whether
the pagination is per-page or per-resource. Later, we should replace
this extension with two independent OpenAPI options: one_indexed
(defaulting to false) and pagination_basis (defaulting to resource but
can be overridden to page).
If users don't specify a limit, the SDK will include a hard-coded limit
of 100 resources per request. We could add this to the OpenAPI spec as
an option default_limit, which is useful for any non-paginated APIs that
later expose pagination options and allow the SDK to gracefully support
those. However, we don't want to encourage folks to use this pattern:
all new list APIs are required to be paginated from the start.

## Tests
<!-- How is this tested? -->

---------

Co-authored-by: Xinjie Zheng <xinjie.zheng@databricks.com>