Better Java examples (#200)
## Changes
This PR makes several changes to examples and documentation for the Java
SDK.
1. Add documentation comments for the Wait class and methods, describing
when they are returned and how to use them to retrieve API responses.
2. Check in examples from
https://docs.databricks.com/en/dev-tools/sdk-java.html as a reference
for @PaulCornellDB for maintaining the docs page. This way, examples are
updated as necessary to account for any changes in the SDK.
3. Refreshed all existing examples to work with the current version of
the Java SDK.
4. Refactored all non-Spring Boot examples into a single maven project
in `examples/docs` so that users can find all relevant examples in a
single directory
(examples/docs/src/main/java/com/databricks/sdk/examples). This also
means that there are fewer `pom.xml` files that need to be maintained.
5. Bump the version of the Java SDK used in the example maven projects
as part of the release flow.

Later, as part of the release process, we may need to set the `SNAPSHOT`
version to make it possible to test local changes to the SDK more
easily.

## Tests
- [x] Compilation of the maven projects in the `examples/` directory
passes.
- [ ] Unable to test the release workflow (as that forcibly resets to
the last commit on main).