Add basic support for HTTP proxies (#241)
## Changes
This PR sets `useSystemProperties()` when building the Commons HTTP
client. This allows users to configure an HTTP proxy by setting the
`https.proxyHost` and `https.proxyPort` parameters (see
https://docs.oracle.com/javase/6/docs/technotes/guides/net/proxies.html).

Closes #111.

## Tests
Added an example. Started the HTTP proxy from
https://github.com/databricks/databricks-sdk-go/pull/825, created a run
configuration for the new example that sets `https.proxyHost` to
`localhost` and `https.proxyPort` to `8443`, and run. I see the proxy
server handling a request and the example completes successfully.