Fix one-shot list APIs to not return null (#266)
## Changes
Currently, non-paginated list APIs in the Java SDK simply return the
field in the response with the listed items. However, when there are no
items in the collection being listed. This PR changes all list APIs to
return a Paginator so that the API response is never `null`.

Resolves #197.

## Tests
Tested listing clusters using an SP with no access to any clusters,
which passed:
```
10:12 [DEBUG] > GET /api/2.0/clusters/list
< 200 OK
< { }
```