[PECO-1542] Add a way to provide proxy details to SDK (#271)
## Changes
<!-- Summary of your changes that are easy to understand -->
Duplicate of #242. Re-created to sign commits. (This PR adds a way to
provide proxy configs in the SDK directly, ability to pick up the system
properties and authentication for proxy via basic or negotiate-kerberos
schemes)

## Tests
<!-- How is this tested? -->
Documented in #242 (These changes were tested in an environment
consisting of a proxy server with kerberos authentication hosted on
Ubuntu and Windows VMs. The detailed testing details are documented
here:
https://docs.google.com/document/d/1Zxlfx-R_JytFqMZfQMBfSQ8mj4W-aHa60vw4KdjUkEc/edit)