Fix interrupt handling (#272)
## Changes
Current implementation does not work well when client interrupts a
thread where SDK code is running. Once interrupted, the logic continues
to run till the timeout (default 20mins) without sleeping at all. Each
time thread.sleep is called, it throws interrupt exception and this
results in calling the status API continuously without delay.

The PR is changing the handling to throw DatabricksException when the
thread is interrupted so that clients can stop the processing and handle
the exception as they wish.

## Tests
All unit tests are successful.