Fix OIDC Endpoint Fetching in DatabricksConfig for Workspace Clients (#277)
## Changes
The logic for determining OIDC endpoints depends on whether a user is
authenticating to the workspace or account. The Java SDK uses presence
of account ID to make this decision, which is actually incorrect: logic
for determining whether a client is configured at the workspace or
account-level is based on the hostname. This PR fixes this logic to
match the Python SDK. As a result, users are able to set the account ID
configuration for Workspace clients and will still be able to
authenticate with OAuth.

## Tests
- [x] Some unit tests added to verify this behavior.
- [x] Manually verified the old behavior, and verified that this doesn't
recur post-fix.