Support partners in SDK (#291)
## Changes
This PR is a port of
https://github.com/databricks/databricks-sdk-go/pull/925 to the Java
SDK.

Partners of Databricks need a mechanism to register themselves in
libraries or applications that they write. In this way, requests made by
users of those libraries will include sufficient information to link
those requests to the original users.

This PR adds a new function `withPartner` that can be used by a partner
to add partner information to the User-Agent header for requests made by
the SDK. The new header will have the form `partner/<parther id>`. The
partner identifier is opaque for the SDK, but it must be alphanumeric.

This PR also lifts the requirement that there may only be a single copy
of an entry in the user agent. This allows multiple partners to register
themselves in the same library or application.

Additionally, this PR adds validation on keys and values of entries in
the user agent to match the Go SDK: alphanumeric keys and alphanumeric
or semver values.

## Tests
Unit tests cover the implementation of the `UserAgent` class.