Accept any `HttpRequest` instead of `HttpUriRequest` in `CommonHttpClient` (#305)
## Changes

This PR addresses a regression reportedly introduced in PR #290 when
connecting via a proxy that requires authentication (see context).

The problem comes from a casting exception in `CommonHttpClient` which
attempts to cast a `BasicHttpRequest` (which represents the next request
after receiving the auth required response) into a `HttpUriRequest`.
This PR solves the casting issue by processing the ancestor of both
classes, `HttpRequest`.

### Context

The regression was detected when trying to authenticate via a Proxy that
requires authentication. In particular, the request to respond to the
proxy's authentication request is automatically created as a
`BasicHttpRequest` by the apache `HttpClient`.

## Tests

Unit tests and integration tests are passing. The fix was also verified
by the user who uncovered the issue.

Note: we should add a regression test that simulate the proxy setting.
However, I'd like to take the time to experiment with different
configurations. Given that the change is relatively simple, I'd
recommend to proceed with this PR and add the test in a follow-up PR.