[Feature] Add credential provider for Azure Github OIDC (#307)
## Changes

This PR adds a `CredentialsProvider` to authenticate with Azure from a
Github action using OIDC. The implementation echoes the one in the Go
and Python SDKs.

## Tests

Most of the new code paths are covered by the test. Note that I'm adding
the `org.json` library to help building JSON object in a concise way.
The library is fairly mature and part of the Android SDK.