[Feature] Experimental: Make retry strategy configurable (#363)
Adds support to configure Retry Strategy in HttpClient. Currently only
the default retry strategy is used. This strategy retries 3 times and
does not have any sleep interval in between. For our use case, we would
prefer using the ExponentialBackOffStrategy.

This does not affect the default behavior but gives additional options
to the users.