[Fix] Correctly include query parameters for APIs whose request objects contain the body as a field (#401)
## What changes are proposed in this pull request?
The Databricks API spec has two different styles of APIs, those where
the request body is a field in the `Request` structure, and those where
the request body is said structure (not counting the parameters passed
via the URI path and query string). The former style of APIs was
incorrectly implemented, preventing such APIs from using any query
parameters.

Generally, the approach taken here is to move the construction of the
request object into the autogenerated `Impl` classes and providing a
fully materialized `Request` object to the `ApiClient`. To do so, we
expose the `<T> T execute(Request in, Target Class<T>)` method as
public.

Separately, I've removed the per-HTTP-method methods from `ApiClient`,
opting instead to use the singular `execute`, as they purely create more
maintenance burden for us and decrease the surface area we need to
maintain.

## How is this tested?
Added `AppsImplTest` to verify that the no_compute flag is included in
the request passed to the `ApiClient`.