Add changelog that needs to added in v0.55.0 (#486)
## What changes are proposed in this pull request?

Add the changelog that needs to be added in v0.55.0. We forgot that the
release v0.55.0 failed and tried to release v0.56.0, and it also failed.
The current state of main contains the changes from both v0.55.0 and
v0.56.0. So, adding the complete changelog for the v0.55.0 release.

## How is this tested?

Existing CI.