Add types for all the Databricks error detail types. (#497)
## What changes are proposed in this pull request?

This PR is a first step in adding support for all the Databricks error
detail types. Specifically, this PR introduces AutoValue classes for
each error detail and make sure that these are parsed properly. However,
the types are not exposed yet in the actual DatabricksError exception,
this will come in a subsequent PR.

While it might not look as such, the change in `ApiErrorBody` is
practically a no-op because:

1. The server can only return one error info.
2. Only error info types were extracted from the list of details.

## How is this tested?

Unit tests focused on deserialization.

NO_CHANGELOG=true