[PECOBLR-15] Respect OAuth Discovery URL for ExternalBrowserCredentialsStrategy (#505)
## What changes are proposed in this pull request?

- **WHAT** : Discovery URL is currently being honored only in the M2M
flow. This PR makes changes to honor it for U2M and other flows as well.
Discovery URL was added by me last year in SDK for M2M flow
[Link](https://github.com/databricks/databricks-sdk-java/pull/336)
- ~~I have also added a fallback to defaultOIDC flow to make this
backward compatible too.~~ skipped the fallback in this PR, will raise
in a followup if I get approval from the SDK team
- **WHY** : A major customer ask is to extend discovery URL for U2M flow
as well.

## How is this tested?

Tested that discovery URL is being honored in U2M as well. Tested this
using the OSS JDBC with the following URL
1. Success flow test : 
```
String jdbcUrl =
        "jdbc:databricks://e2-dogfood.staging.cloud.databricks.com:443/default;transportMode=http;ssl=1;httpPath=/sql/1.0/warehouses/dd43ee29fedd958d;"
            + "AuthMech=11;Auth_Flow=2;LogLevel=6;UseThriftClient=0;EnableTokenCache=0;OIDCDiscoveryEndpoint=https://e2-dogfood.staging.cloud.databricks.com/oidc/.well-known/oauth-authorization-server";
```
2. Failure flow test : 
```
String jdbcUrl =
        "jdbc:databricks://e2-dogfood.staging.cloud.databricks.com:443/default;transportMode=http;ssl=1;httpPath=/sql/1.0/warehouses/dd43ee29fedd958d;"
            + "AuthMech=11;Auth_Flow=2;LogLevel=6;UseThriftClient=0;EnableTokenCache=0;OIDCDiscoveryEndpoint=https://google.com";
```