Update SDK to latest API definition (#511)
## What changes are proposed in this pull request?

Update SDK to the latest API definition.

This PR also adds a cleanup script that selectively opts out handwritten
files in the service directory, which is currently deleted completely.
This is needed as we are adding handwritten code in the service
directory.

Note for reviewers: I have manually added few files in iam and sharing
package for backward compatibility.

## How is this tested?

Existing CI.