Add LroOptions for Long Running Operation (#519)
## 🥞 Stacked PR
Use this
[link](https://github.com/databricks/databricks-sdk-java/pull/519/files)
to review incremental changes.
-
[**stack/add-lro-options**](https://github.com/databricks/databricks-sdk-java/pull/519)
[[Files
changed](https://github.com/databricks/databricks-sdk-java/pull/519/files)]

---------
## What changes are proposed in this pull request?

This PR adds LroOptions for Long-Running Operations. These options are
similar to those in the Go SDK. Currently, this only has a timeout, but
this can be extended in the future.
 
## How is this tested?

N/A

NO_CHANGELOG=true