Add `DATABRICKS_DISABLE_RETRIES` config option to disable the default retry mechanism. (#523)
## What changes are proposed in this pull request?

This PR adds a config-level property to disable the default retry
mechanism. That is, no request will be retried no matter its error
code/status. This is implemented via a new retry strategy
`NoRetryStrategy` which systematically returns `false`.

This PR addresses some of the limitations called out in issue: #289. 

**Why not providing control on the number of retries directly?**

The number of retries should be considered an implementation detail of a
specific retry strategy. The Databricks SDKs are meant to evolve to a
model where the retry conditions are more and more server controlled
(e.g. leverage the RetryInfo error details). Rather than providing
control over the number of retries, we intent to provide users with
control over (i) the overall timeout of the method call (including
retries), and (ii) the retry strategy used on a specific client.

## How is this tested?

Unit + Integration tests.