Add Well Known Types serialization utils (#518)
## What changes are proposed in this pull request?
Add Well Known Types serialization utils. This PR configures the default
ObjectMapper for the Java SDK to serialize the Google Well Known Types:
https://protobuf.dev/reference/protobuf/google.protobuf/

For Duration and Timestamp, we use the official protojson serialization.
For FieldMask, we do introduce a custom serialization since the Google
FieldMask serialization is not aligned with the expected server values.

## How is this tested?
Added unit tests.

NO_CHANGELOG=true