Fixed configure method for non-refreshable tokens (#526)
## What changes are proposed in this pull request?

We recently added support for disabling auto token refresh
(https://github.com/databricks/databricks-sdk-java/pull/525). The
current code is hardcoded to handle refreshable tokens, and it
forcefully refreshes the token on every call of `configure()`
This PR proposed using a token (no matter if it is refreshable or not)
if it is valid. Once the ExternalBrowserCredentialsProvider is run, it
stores the token on disk. If the application is run again, it checks the
token and tries to use it.

## How is this tested?

Added Unit tests

NO_CHANGELOG=true