Add Support for Field Mask in query parameters. (#535)
## What changes are proposed in this pull request?
Add Support for Field Mask in query parameters. 
We are introducing a rich type for FieldMask. This PR ensures that
FieldMasks are serialized correctly when used as Query parameters (which
is the case for all non batch operations).

Support for FieldMask is added as part of generation framework in a
separate PR.

More information about field mask:
https://developers.google.com/workspace/slides/api/guides/field-masks

Other Well Known types are not currently supported as query parameters.
Tests are being generated, so they are not part of this PR.

## How is this tested?
Chicken and egg problem. We have 3 components, all of which are needed
to test this:
1. New structures with the new types.
2. Generated tests.
3. The change in this PR.
Components 1 and 2 are generated and part of another PR. We cannot test
this without those components. But at the same time, those components
need this change or the generated tests will fail.
I generated locally the SDK with all 3 changes at the same time and run
`make test` (diff includes many changes. Releavant changes in
HttpPathTest.java):


https://github.com/databricks/databricks-sdk-java/pull/536/files#diff-e0a2f237fe2b016080aeb672d5793f68486a7b897c7252480c7c00a76f08796f

NO_CHANGELOG=true